Move SettingsLoader to a better place
